// packages/shared/src/api/types.ts
export interface SearchParams {
  entity_name?: string;
  department?: string;
  city?: string;
  phase?: string;
  modality?: string;
  min_value?: number;
  max_value?: number;
  from_date?: string;
  to_date?: string;
  status?: string;
  limit?: number;
  query?: string;
  search_mode?: 'all_words' | 'phrase' | 'any_word';
}